Suggestions for Chapter 1 (#1)
* Suggestions for Chapter 1

Mostly reporting grammar typos. If a change is made not related to grammar, and it is not clear why I made the change, it is probably a stylistic choice. If I see something that could be changed but I do not know how I would change it, I provide a comment in the file.

General comments:
the contraction i.e. is followed by a ",". This is both grammatically correct (i checked a few dictionaries online) and latex behaves correctly and makes a smaller space between the words. "i.e. smth smth" is interpreted by the compiler as the end of a sentence and the start of a new sentence, meanwhile "i.e., smth smth" is one sentence.

* Update hm.tex

Co-authored-by: Marcello Seri <mseri@users.noreply.github.com>

* Update hm.tex

Co-authored-by: Marcello Seri <mseri@users.noreply.github.com>

* Update hm.tex

Co-authored-by: Marcello Seri <mseri@users.noreply.github.com>

* Update hm.tex

Co-authored-by: Marcello Seri <mseri@users.noreply.github.com>

* Update hm.tex

Co-authored-by: Marcello Seri <mseri@users.noreply.github.com>

* Update hm.tex

Co-authored-by: Marcello Seri <mseri@users.noreply.github.com>

* Update hm.tex

Co-authored-by: Marcello Seri <mseri@users.noreply.github.com>

* Update hm.tex

Co-authored-by: Marcello Seri <mseri@users.noreply.github.com>

* Update hm.tex

Co-authored-by: Marcello Seri <mseri@users.noreply.github.com>

* Update hm.tex

Co-authored-by: Marcello Seri <mseri@users.noreply.github.com>